Frontend(navbar and search bar): Adjusted the header component to not show up on login page.
