+++ onemed1a-frontend/src/app/layout.js
@@ -1,8 +1,7 @@
 import { Poppins } from "next/font/google";
 import "./globals.css";
-import Logo from "@/components/Logo";
-import MediaNavigation from "@/components/MediaNavigation";
 import PropTypes from "prop-types";
+import ConditionalAppChrome from "../components/ConditionalAppChrome"; // [`ConditionalAppChrome`](onemed1a-frontend/src/components/ConditionalAppChrome.js)
 
 const poppins = Poppins({
   subsets: ["latin"],
@@ -22,28 +21,9 @@ export default function RootLayout({ children }) {
         <meta charSet="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
       </head>
-
       <body className={`${poppins.variable} antialiased bg-white text-neutral-900`}>
-        {/* Sticky logo only */}
-        <header
-          className="sticky top-0 z-50 bg-white/90 backdrop-blur supports-[backdrop-filter]:bg-white/70 border-b border-neutral-200 shadow-sm"
-          aria-label="Application Header"
-        >
-          <div className="mx-auto max-w-7xl px-4 py-4 flex justify-center">
-            <Logo />
-          </div>
-        </header>
-
-        {/* Non-sticky navigation (no extra logo here) */}
-        <nav className="bg-white border-b border-neutral-200" aria-label="Media categories">
-          <div className="mx-auto max-w-7xl px-4 py-3 flex justify-center">
-            <MediaNavigation />
-          </div>
-        </nav>
-
-        <main id="main" className="min-h-screen">
-          {children}
-        </main>
+        <ConditionalAppChrome />
+        {children}
       </body>
     </html>
   );
