+++ onemed1a-frontend/src/app/page.js
@@ -1,50 +1,10 @@
-'use client';
+import { cookies } from 'next/headers';
+import { redirect } from 'next/navigation';
 
-import Link from 'next/link';
-import { useState } from 'react';
-import { useRouter } from 'next/navigation';
-import { checkAccountDetails } from '@/api/userAPI';
-
-export default function LoginPage() {
-  const router = useRouter();
-  const [email, setEmail] = useState(''), [password, setPassword] = useState(''), [error, setError] = useState('');
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();            // <-- stop the page reload
-    setError('');
-    try {
-      const user = await checkAccountDetails({ email, password });
-      console.log(user)
-      document.cookie = `userId=${encodeURIComponent(user.id)}; Path=/; Max-Age=2592000; SameSite=Lax`;
-      router.push('/movie');
-    } catch {
-      setError('Login failed');
-    }
-  };
-
-  return (
-      <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-        <div className="max-w-md w-full space-y-8">
-          <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">Sign in to your account</h2>
-          <form className="mt-8 space-y-6" onSubmit={handleSubmit}>
-            <div className="space-y-4">
-              <input id="email" type="email" value={email} onChange={e=>setEmail(e.target.value)} required
-                     className="appearance-none rounded-md block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" placeholder="Email address" />
-              <input id="password" type="password" value={password} onChange={e=>setPassword(e.target.value)} required
-                     className="appearance-none rounded-md block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" placeholder="Password" />
-            </div>
-
-            {error && <p className="text-red-500 text-sm">{error}</p>}
-
-            <button type="submit" className="w-full py-2 px-4 text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none">
-              Sign in
-            </button>
-
-            <div className="text-center text-sm text-gray-600">
-              Don&apos;t have an account? <Link href="/signup" className="font-medium text-indigo-600 hover:text-indigo-500">Sign up</Link>
-            </div>
-          </form>
-        </div>
-      </div>
-  );
+export default function HomePage() {
+  const userId = cookies().get('userId')?.value;
+  if (!userId) {
+    redirect('/login');
+  }
+  redirect('/movie');
 }
